Change license of the software
The new license of the project, named "The nWT License," is an
extremely permissive software license contract which only exists to
keep the copyrights of the software (given the lack of rights to give
anything to the public-domain in France).

The terms are simple and can be understood by anyone, it essentially
requires keeping the copyright notice with the distributed software.

It should be OSI-friendly (Open Source Initiative) and GPL-compatible,
but I haven't taken any action to make it recognized, that's too much
work for the kind of crap intended to be in the nWT umbrella.

Lastly, this commit also updates the copyright notice in the web-site
code and in Fluid Layout, to reflect our ideas.