Introduce Space.css and its default scale of 0.25
Space.css is a configurable spacing system directly encoded in CSS
thanks to CSS custom properties.  It provides functional CSS classes to
control the margins, paddings, and gaps of elements, e.g. m-3.

By default, the scale is provided from rank 0 to 100, and uses 0.25rem
as a base.  Each element in the scale is the simple product between its
rank and the base:

    .m-x { x*base }

The base can be changed by overriding the --nwt-space custom property
in user's :root {}.

The expected way of using Space CSS is configuring the base and copying
the classes you will use from the scale, as the entire thing is 1200
CSS lines, most of which will never be used by anyone (does someone
really wants rank 73 of padding?).

Similarly to Fluid Layout, Space CSS is distributed under the nWT
License terms.